Fixed linux support
