+++ .vscode/settings.json
@@ -1,2 +1,11 @@
 {
+    "cSpell.words": [
+        "Codeuctivity",
+        "PDFs",
+        "Rasterize",
+        "Rasterizer",
+        "Rasterizes",
+        "Xunit",
+        "pngs"
+    ]
 }
