+++ PdfjsSharp/PdfjsSharp.csproj
@@ -8,10 +8,10 @@
     <PackageTags>Pdf Png</PackageTags>
     <Authors>Stefan Seeland</Authors>
     <Company>Codeuctivity</Company>
-    <AssemblyVersion>1.0.6.0</AssemblyVersion>
-    <FileVersion>1.0.6.0</FileVersion>
-    <Version>1.0.6</Version>
-    <PackageVersion>1.0.6-prerelease</PackageVersion>
+    <AssemblyVersion>1.0.7.0</AssemblyVersion>
+    <FileVersion>1.0.7.0</FileVersion>
+    <Version>1.0.7</Version>
+    <PackageVersion>1.0.7-prerelease</PackageVersion>
     <PackageIcon>NugetIcon.png</PackageIcon>
     <PackageProjectUrl>https://github.com/Codeuctivity/PdfjsSharp</PackageProjectUrl>
     <Description>Pdfjs for .net</Description>
@@ -66,6 +66,7 @@
   <ItemGroup>
     <EmbeddedResource Include="Rasterize.js" />
     <EmbeddedResource Include="node_modules.win.zip" />
+    <EmbeddedResource Include="node_modules.linux.zip" />
   </ItemGroup>
 
   <ItemGroup>
